+++ generators/client/files.js
@@ -2,6 +2,7 @@ const jhipsterConstants = require('generator-jhipster/generators/generator-const
 
 const ANGULAR_DIR = jhipsterConstants.ANGULAR_DIR;
 const REACT_DIR = jhipsterConstants.REACT_DIR;
+const VUE_DIR = jhipsterConstants.VUE_DIR;
 const TEST_DIR = jhipsterConstants.CLIENT_TEST_SRC_DIR;
 
 const clientFiles = {
@@ -9,7 +10,6 @@ const clientFiles = {
         {
             condition: generator => generator.clientFramework === 'angularX',
             templates: [
-                { file: 'angular/tsconfig-aot.json', renameTo: () => 'tsconfig-aot.json' },
                 { file: 'angular/tsconfig.json', renameTo: () => 'tsconfig.json' },
                 { file: 'angular/package.json', renameTo: () => 'package.json' },
                 { file: 'angular/eslintignore', renameTo: () => '.eslintignore' }
@@ -20,8 +20,15 @@ const clientFiles = {
             templates: [
                 { file: 'react/tsconfig.json', renameTo: () => 'tsconfig.json' },
                 { file: 'react/package.json', renameTo: () => 'package.json' },
-                { file: 'react/eslintrc.json', renameTo: () => '.eslintrc.json' },
                 { file: 'react/eslintignore', renameTo: () => '.eslintignore' }
+            ]
+        },
+        {
+            condition: generator => generator.clientFramework === 'vue',
+            templates: [
+                { file: 'vue/tsconfig.json', renameTo: () => 'tsconfig.json' },
+                { file: 'vue/package.json', renameTo: () => 'package.json' },
+                { file: 'vue/eslintignore', renameTo: () => '.eslintignore' }
             ]
         }
     ],
@@ -74,15 +81,46 @@ const clientFiles = {
             ]
         }
     ],
+    vueMain: [
+        {
+            condition: generator => generator.clientFramework === 'vue',
+            templates: [
+                {
+                    file: 'vue/home/home.vue',
+                    renameTo: () => `${VUE_DIR}core/home/home.vue`
+                },
+                {
+                    file: 'vue/docs/docs.vue',
+                    renameTo: () => `${VUE_DIR}admin/docs/docs.vue`
+                },
+                {
+                    file: 'vue/jhi-navbar/jhi-navbar.vue',
+                    renameTo: () => `${VUE_DIR}core/jhi-navbar/jhi-navbar.vue`
+                },
+                {
+                    file: 'vue/account/account.service.ts',
+                    renameTo: () => `${VUE_DIR}account/account.service.ts`
+                },
+                {
+                    file: 'vue/account/account.service.spec.ts',
+                    renameTo: () => `${TEST_DIR}spec/app/account/account.service.spec.ts`
+                }
+            ]
+        },
+        {
+            condition: generator => generator.clientFramework === 'vue' && !generator.skipUserManagement,
+            templates: [
+                {
+                    file: 'vue/user-management/user-management.component.ts',
+                    renameTo: () => `${VUE_DIR}admin/user-management/user-management.component.ts`
+                }
+            ]
+        }
+    ],
     testE2EWithMain: [
         {
             condition: generator => generator.clientFramework === 'react',
             templates: [
-                {
-                    file: 'react/shared/layout/header/header.spec.tsx',
-                    method: 'processJsx',
-                    renameTo: () => `${TEST_DIR}spec/app/shared/layout/header/header.spec.tsx`
-                },
                 {
                     file: 'react/test/protractor.conf.js',
                     renameTo: () => `${TEST_DIR}protractor.conf.js`
@@ -90,6 +128,10 @@ const clientFiles = {
                 {
                     file: 'react/test/e2e/administration.spec.ts',
                     renameTo: () => `${TEST_DIR}e2e/modules/administration/administration.spec.ts`
+                },
+                {
+                    file: 'react/test/e2e/account.spec.ts',
+                    renameTo: () => `${TEST_DIR}e2e/modules/account/account.spec.ts`
                 }
             ]
         },
@@ -103,6 +145,23 @@ const clientFiles = {
                 {
                     file: 'angular/test/e2e/administration.spec.ts',
                     renameTo: () => `${TEST_DIR}e2e/admin/administration.spec.ts`
+                }
+            ]
+        },
+        {
+            condition: generator => generator.clientFramework === 'vue',
+            templates: [
+                {
+                    file: 'vue/test/protractor.conf.js',
+                    renameTo: () => `${TEST_DIR}protractor.conf.js`
+                },
+                {
+                    file: 'vue/test/e2e/administration.spec.ts',
+                    renameTo: () => `${TEST_DIR}e2e/modules/administration/administration.spec.ts`
+                },
+                {
+                    file: 'vue/test/e2e/account.spec.ts',
+                    renameTo: () => `${TEST_DIR}e2e/modules/account/account.spec.ts`
                 }
             ]
         }