+++ generators/client/index.js
@@ -7,7 +7,7 @@ const writeFiles = require('./files').writeFiles;
 
 module.exports = class extends ClientGenerator {
     constructor(args, opts) {
-        super(args, Object.assign({ fromBlueprint: true }, opts)); // fromBlueprint variable is important
+        super(args, { fromBlueprint: true, ...opts }); // fromBlueprint variable is important
 
         const jhContext = (this.jhipsterContext = this.options.jhipsterContext);
 
@@ -16,8 +16,6 @@ module.exports = class extends ClientGenerator {
         }
 
         this.configOptions = jhContext.configOptions || {};
-        // This sets up options for this sub generator and is being reused from JHipster
-        jhContext.setupClientOptions(this, jhContext);
 
         // This adds support for a `--skip-i18n` flag for unit test
         this.option('skip-i18n', {
@@ -36,7 +34,7 @@ module.exports = class extends ClientGenerator {
                 // this.packagejs = jhipsterPackagejs;
             }
         };
-        return Object.assign(initPhaseFromJHipster, initNodeClientPhaseSteps);
+        return { ...initPhaseFromJHipster, ...initNodeClientPhaseSteps };
     }
 
     get prompting() {
@@ -44,12 +42,11 @@ module.exports = class extends ClientGenerator {
     }
 
     get configuring() {
-        // Here we are not overriding this phase and hence its being handled by JHipster
         return super._configuring();
     }
 
-    get default() {
-        const defaultPhaseFromJHipster = super._default();
+    get composing() {
+        const defaultPhaseFromJHipster = super._composing();
         // disable i18n
         if (this.options['skip-i18n']) {
             defaultPhaseFromJHipster.composeLanguages = {};
@@ -57,15 +54,34 @@ module.exports = class extends ClientGenerator {
         return defaultPhaseFromJHipster;
     }
 
+    get loading() {
+        return this._loading();
+    }
+
+    get preparing() {
+        return this._preparing();
+    }
+
+    get default() {
+        const defaultPhaseFromJHipster = super._default();
+        const defaultNodeClientPhaseSteps = {
+            // variables to use in templates
+            setupCustomClientConsts() {
+                this.protractorTests = true;
+            }
+        };
+        return { ...defaultPhaseFromJHipster, ...defaultNodeClientPhaseSteps };
+    }
+
     get writing() {
         const phaseFromJHipster = super._writing();
         const jhipsterNodeClientPhaseSteps = writeFiles();
-        return Object.assign(phaseFromJHipster, jhipsterNodeClientPhaseSteps);
+        return { ...phaseFromJHipster, ...jhipsterNodeClientPhaseSteps };
     }
 
-    get install() {
-        // Here we are not overriding this phase and hence its being handled by JHipster
-        return super._install();
+    get postWriting() {
+        // you can use this phase to overwrite package.json
+        return this._postWriting();
     }
 
     get end() {