+++ generators/entity-client/index.js
@@ -1,0 +1,29 @@
+/* eslint-disable consistent-return */
+const chalk = require('chalk');
+const EntityClientGenerator = require('generator-jhipster/generators/entity-client');
+const writeFiles = require('./files').writeFiles;
+
+module.exports = class extends EntityClientGenerator {
+    constructor(args, opts) {
+        super(args, { fromBlueprint: true, ...opts }); // fromBlueprint variable is important
+
+        if (!this.jhipsterContext) {
+            this.error(`This is a JHipster blueprint and should be used only like ${chalk.yellow('jhipster --blueprints nodejs')}`);
+        }
+    }
+
+    get default() {
+        // Here we are not overriding this phase and hence its being handled by JHipster
+        return super._default();
+    }
+
+    get writing() {
+        const phaseFromJHipster = super._writing();
+        const jhipsterNodeEntityClPhaseSteps = writeFiles();
+        return { ...phaseFromJHipster, ...jhipsterNodeEntityClPhaseSteps };
+    }
+
+    get postWriting() {
+        return this._postWriting();
+    }
+};