+++ test/server.spec.js
@@ -10,10 +10,11 @@ function getPreCondition() {
     return helpers
         .run('generator-jhipster/generators/server')
         .withOptions({
-            'from-cli': true,
+            fromCli: true,
             skipInstall: true,
             blueprints: 'nodejs',
-            skipChecks: true
+            skipChecks: true,
+            'skip-i18n': true
         })
         .withGenerators([
             [
@@ -29,6 +30,7 @@ function commonAssertion() {
     assert.file(`${SERVER_NODEJS_DIR}e2e/app.e2e-spec.ts`);
     assert.file(`${SERVER_NODEJS_DIR}e2e/user.e2e-spec.ts`);
     assert.file(`${SERVER_NODEJS_DIR}e2e/jest.e2e.config.json`);
+    assert.file(`${SERVER_NODEJS_DIR}e2e/setup.test.js`);
     assert.file('src/main/docker/app.yml');
     assert.file('src/main/docker/mysql.yml');
     assert.noFile('src/main/docker/mongodb.yml');